backend/requirements.txt
# backend/requirements.txt

Flask
python-dotenv
PyGithub
langchain
langchain-google-genai
google-generativeai
PyPDF2
flask_cors
langchain-text-splitters
faiss-cpu
langchain-community

